requirements2.txt
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
django-celery-beat==1.1.0
django-celery-results==1.0.1
djangorestframework==3.7.7
ephem==3.7.6.0
html5lib==1.0.1
idna==2.6
kombu==4.1.0
lxml==4.1.1
numpy==1.14.0
pandas==0.22.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
